Extract some private submodules from 'bat::assets'
