+++ src/assets/assets_metadata.rs
