+++ src/lib.rs
@@ -22,9 +22,9 @@
 mod macros;
 
 pub mod assets;
-pub mod assets_metadata;
-#[cfg(feature = "build-assets")]
-mod build_assets;
+pub mod assets_metadata {
+    pub use super::assets::assets_metadata::*;
+}
 pub mod config;
 pub mod controller;
 mod decorations;
